+++ benchmarks/verus/verina/files/verina_advanced_44_task.rs
@@ -8,7 +8,7 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn max_subarray_sum_divisible_by_k(arr: &Vec<i32>, k: i32) -> (result: i32)
+fn max_subarray_sum_divisible_by_k(arr: &Vec<i8>, k: i8) -> (result: i8)
     requires k > 0,
     ensures true,
 // </vc-spec>
