+++ benchmarks/verus/verina/files/verina_advanced_4_task.rs
@@ -11,7 +11,7 @@ verus! {
 fn longest_increasing_subsequence(a: &Vec<i32>) -> (result: i32)
     ensures
         result >= 0,
-        result <= a.len(),
+        result <= a.len() as i32,
 // </vc-spec>
 // <vc-code>
 {
