+++ benchmarks/verus/verina/files/verina_basic_41_task.rs
@@ -8,7 +8,7 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn has_only_one_distinct_element(a: &Vec<i32>) -> (result: bool)
+fn has_only_one_distinct_element(a: &Vec<i8>) -> (result: bool)
     requires a.len() > 0,
     ensures
         result ==> (forall|i: int, j: int| 0 <= i < a.len() && 0 <= j < a.len() ==> a[i] == a[j]),
