+++ benchmarks/verus/verina/files/verina_basic_83_task.rs
@@ -8,11 +8,11 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn concat(a: &Vec<i32>, b: &Vec<i32>) -> (result: Vec<i32>)
+fn concat(a: &Vec<i8>, b: &Vec<i8>) -> (result: Vec<i8>)
     ensures
         result.len() == a.len() + b.len(),
-        forall|k: int| 0 <= k < a.len() ==> result[k] == a[k],
-        forall|k: int| 0 <= k < b.len() ==> result[k + a.len()] == b[k],
+        forall|k: int| 0 <= k < a.len() ==> result[k] as i8 == a[k],
+        forall|k: int| 0 <= k < b.len() ==> result[k + a.len()] as i8 == b[k],
 // </vc-spec>
 // <vc-code>
 {
